+++ src/Context/PackageRules.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ItpContext\Context;
+
+use ItpContext\Contract\RuleIdentifier;
+
+enum PackageRules implements RuleIdentifier
+{
+    case FrameworkAgnostic;
+    case CatalogByConvention;
+    case TokenFirstDiscovery;
+    case AgentFriendlyMarkdown;
+}